Removing different stages for continuity
According to [the Travis-CI documentation](https://docs.travis-ci.com/user/build-stages/#data-persistence-between-stages-and-jobs), data is not shared between stages on the CI system itself and requires the use for an external storage provider. In this case, we can live with one single stage.

Also added the `psalm.xml.dist` configuration for static analysis since it failed the preparation phase.